refactor: introduce TaskEditorViewModel for shared task editing logic
The `AddTaskViewModel` now inherits from `TaskEditorViewModel`, removing redundant code for handling title, description, date, priority, and category selection.
